chore: refactor to use playwright for e2e testing
