+++ e2e/jest.playwright.setup.ts
@@ -1,6 +1,6 @@
 import { Config } from '@jest/types';
 
-import { startServer } from './puppeteer';
+import { startServer } from './playwright';
 
 export default async (globalConfig: Config.GlobalConfig) => {
   await startServer(globalConfig);
