+++ e2e/package.json
@@ -7,17 +7,14 @@
   "dependencies": {
     "@babel/runtime": "^7.10.2",
     "@remirror/core": "^0.11.0",
-    "@types/expect-puppeteer": "^4.4.3",
-    "@types/jest-environment-puppeteer": "^4.3.1",
     "@types/puppeteer": "^3.0.0",
     "@types/wait-on": "^4.0.0",
-    "expect-puppeteer": "^4.4.0",
+    "expect-playwright": "^0.2.4",
     "jest-dev-server": "^4.4.0",
-    "jest-environment-puppeteer": "^4.4.0",
-    "jest-puppeteer": "^4.4.0",
+    "jest-playwright-preset": "^0.2.5",
     "looks-same": "^7.2.3",
-    "pptr-testing-library": "^0.6.0",
-    "puppeteer": "^4.0.0",
+    "playwright": "^1.1.1",
+    "playwright-testing-library": "^1.2.1",
     "signal-exit": "^3.0.3",
     "test-keyboard": "^0.7.6"
   }
